Move TesseractHelper into Text-Grab.Core.Windows, split out the downloader
Batch 4b of the Core split. TesseractHelper.cs was blocked only on
AutomationProfile via TempImagePath() - that landed in Text-Grab.Core in
93c1729, so this batch was purely mechanical:

- TesseractHelper: dropped the cached `Settings DefaultSettings` field
  (AppUtilities.TextGrabSettings, the app's internal settings type) and
  read SettingsAccess.Current inline in GetTesseractPath() instead, same
  shape as every other moved settings touchpoint - TesseractPath and
  Save() are already on ITextGrabSettings, no interface change needed.
  AutomationProfile.Current/.GetTemporaryDirectory() in TempImagePath()
  need no using - same namespace (Text_Grab.Utilities), same as the app.
- TesseractGitHubFileDownloader (the second class in the same file) was
  fully portable - HttpClient, File I/O, no Windows types - so it split
  out to Text-Grab.Core/Utilities/TesseractGitHubFileDownloader.cs rather
  than riding along in Core.Windows. Namespace kept as Text_Grab.Utilities
  per invariant 1; its two app call sites (LanguageSettings.xaml.cs) and
  the test file (OcrTests.cs) need no changes since the app already
  references Core.Windows -> Core transitively.

Added CliWrap to Text-Grab.Core.Windows.csproj. Left CliWrap in
Text-Grab.csproj unchanged - QuickSimpleLookup.xaml.cs still uses it
directly.

Self-assembly-reflection audit: grepped both files for
GetExecutingAssembly/GetEntryAssembly/GetCallingAssembly, typeof(x)
.Assembly, Assembly.GetName/.Location/.CodeBase, and AssemblyMetadata -
zero matches. TempImagePath()'s System.AppContext.BaseDirectory fallback
(used only when AutomationProfile.Current is null) is process entry-point
based, not declaring-assembly based, so it means the same thing after the
move - same reasoning as the LocalFileSettingsProvider check in 93c1729.

No files deferred from this batch's list; TesseractHelper moved cleanly
and the downloader split cleanly.

Gate: dotnet build Tests/Tests.csproj -c Debug -p:Platform=x64 - clean
(0 errors; pre-existing MSB3277 assembly-conflict warnings only).
