+++ Text-Grab.Core.Windows/Text-Grab.Core.Windows.csproj
@@ -48,6 +48,8 @@
     <PackageReference Include="Microsoft.WindowsAppSDK.AI" Version="2.4.4" />
     <!-- GDI+ Bitmap barcode reading/writing (BarcodeReader/BarcodeWriter) for BarcodeUtilities. -->
     <PackageReference Include="ZXing.Net.Bindings.Windows.Compatibility" Version="0.16.14" />
+    <!-- Tesseract.exe process invocation for TesseractHelper. -->
+    <PackageReference Include="CliWrap" Version="3.10.5" />
   </ItemGroup>
 
 </Project>
